Installed Tailwind CSS; moved components folder to src/
