+++ src/pages/index.js
@@ -1,6 +1,6 @@
 import * as React from "react"
 
-import Layout from "./components/Layout"
+import Layout from "../components/Layout"
 
 const index = () => {
   return (
